BINDINGS/RUST: moved tests from lib.rs to tests.rs
